+++ src/models/schema.ts
@@ -63,14 +63,6 @@ export class Schema extends Data implements IResourceBase {
   }
 
   /**
-   * Get possible language types for this kind of extension in priority order.
-   * Used to try to rendezvous on a language type that the user might have installed.
-   */
-  languageTypes(): string[] {
-    return getLanguageTypes(this.type);
-  }
-
-  /**
    * Return a file name for this schema.
    */
   fileName(): string {
@@ -90,7 +82,7 @@ export class Schema extends Data implements IResourceBase {
 
   get ccloudUrl(): string {
     if (isCCloud(this)) {
-      return `https://confluent.cloud/environments/${this.environmentId}/schema-registry/schemas/${this.subject}`;
+      return `https://confluent.cloud/environments/${this.environmentId}/stream-governance/schema-registry/data-contracts/${this.subject}`;
     }
     return "";
   }